Rename _context_copied to _context_owner and set it always
We should also set `_context_copied` to `owner` when we initialize
`_context_fields` from an empty set.

Otherwise, when there are two or more attributes using the decorator, we
will deep-copy `_context_fields` on assignment of the second attribute,
which is not necessary, because `_context_fields` is already owned by
us.

As the variable is used to track who owns the field, rename it
apropriately.