Fix operation on multiple inheritance
If multiple ancestors of a class use the `context` mixin, they will all
have `_context_fields` defined.

Inheritance rules result in the `_context_fields` attribute of the first
class being inherited by any descendants.

This is however incorrect: The `_context_fields` attribute of all
parents needs to be merged, otherwise they will be missing from the
context.

The best approach I found for this was to iterate over the parents using
the `mro` function.

This colletion also may need to be done once when `get_context_data` called,
because subclasses might not use the `context` mixin themselves, so the
collection would never run if it was only done as part of
`__set_name__`.